+++ src/main/java/prgrms/marco/be02marbox/domain/exception/handler/UserExceptionHandler.java
@@ -1,12 +1,12 @@
-package prgrms.marco.be02marbox.domain.user.exhandler;
+package prgrms.marco.be02marbox.domain.exception.handler;
+
+import static org.springframework.http.HttpStatus.*;
 
 import java.util.List;
-import java.util.Map;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.support.DefaultMessageSourceResolvable;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.validation.BindingResult;
@@ -14,18 +14,17 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
+import prgrms.marco.be02marbox.domain.exception.dto.ResponseApiError;
 import prgrms.marco.be02marbox.domain.user.controller.UserController;
 import prgrms.marco.be02marbox.domain.user.exception.UserException;
 
 @RestControllerAdvice(assignableTypes = UserController.class)
-public class ExceptionController {
-
-	public static final String MESSAGE = "message";
+public class UserExceptionHandler {
 
 	private final Logger log = LoggerFactory.getLogger(getClass());
 
 	@ExceptionHandler(MethodArgumentNotValidException.class)
-	public ResponseEntity<Map<String, List<String>>> handleClientException(MethodArgumentNotValidException exception) {
+	public ResponseEntity<ResponseApiError> handleClientException(MethodArgumentNotValidException exception) {
 		log.error("ClientException : {0}", exception);
 
 		BindingResult bindingResult = exception.getBindingResult();
@@ -35,25 +34,25 @@ public class ExceptionController {
 			.toList();
 
 		return ResponseEntity
-			.status(HttpStatus.BAD_REQUEST)
-			.body(Map.of(MESSAGE, errors));
+			.status(BAD_REQUEST)
+			.body(new ResponseApiError(errors, BAD_REQUEST.value()));
 	}
 
 	@ExceptionHandler({UserException.class, BadCredentialsException.class})
-	public ResponseEntity<Map<String, String>> handleUserException(Exception exception) {
+	public ResponseEntity<ResponseApiError> handleUserException(Exception exception) {
 		log.error("UserException : {0}", exception);
 
 		return ResponseEntity
-			.status(HttpStatus.BAD_REQUEST)
-			.body(Map.of(MESSAGE, exception.getMessage()));
+			.status(BAD_REQUEST)
+			.body(new ResponseApiError(List.of(exception.getMessage()), BAD_REQUEST.value()));
 	}
 
 	@ExceptionHandler
-	public ResponseEntity<Map<String, String>> handleException(Exception exception) {
+	public ResponseEntity<ResponseApiError> handleException(Exception exception) {
 		log.error("Exception : {0}", exception);
 
 		return ResponseEntity
-			.status(HttpStatus.INTERNAL_SERVER_ERROR)
-			.body(Map.of(MESSAGE, exception.getMessage()));
+			.status(INTERNAL_SERVER_ERROR)
+			.body(new ResponseApiError(List.of(exception.getMessage()), INTERNAL_SERVER_ERROR.value()));
 	}
 }
